update/delete duplicate programme files
