+++ programme/celsius-to-fahrenheit/celsius-to-fahrenheit.c
@@ -1,12 +1,12 @@
-// Write a program to convert the temperature from Celsius to Fahrenheit
+// Write a program to convert the temperature from celsius to fahrenheit
 
 #include <stdio.h>
 
 void main()
 {
-  float c;
-  printf("Enter temp. (in C) : ");
+  float c, f;
+  printf("Input ( C )  : ");
   scanf("%f", &c);
-  float f = ((9 * c) / 5) + 32;
-  printf("Temp. in F : %f\n", f);
+  f = ((9 * c) / 5) + 32;
+  printf("Output ( F ) : %.2f\n", f);
 }
