--- programme/celsius-to-fahrenheit/celsius-to-fahrenheit1.py
@@ -1,8 +1,0 @@
-# Write a program to convert the temperature from Celsius to Fahrenheit
-celsius_1 = float(input("Input ( C )  : "))
-
-# For Converting the temperature to Fahrenheit
-Fahrenheit_1 = (celsius_1 * 1.8) + 32
-
-# print the result
-print('Output ( F ) : %.2f' % (Fahrenheit_1))
