+++ programme/palindrome-number/check_palindrome.py
@@ -1,3 +1,4 @@
+# Write a program to check palindrome number
 def check_palindrome(n: int) -> str:
     """Checks if n is a palindrome"""
     if str(n) == str(n)[::-1]:
